refactor(lib): extract volume-ownership repair subsystem from docker.ts
Move the privileged ownership-repair helpers (repairRootOwnedBindMounts,
repairNamedVolumeOwnership, repairManagedNamedVolumes, and the
SERVICE_NAMED_VOLUMES const) out of the compose driver into a dedicated
control-plane/volume-ownership.ts. These spin up throwaway `alpine chown`
containers and reach into operator-ids/secrets — a distinct privileged
concern from docker.ts's pure compose-command wrapping.

docker.ts's execFile `run` helper is now exported (package-internal, not
surfaced via the barrel) so the repair subsystem reuses it rather than
re-implementing it. Behavior is unchanged; the barrel re-exports the moved
symbols from their new home, and the sole consumer (ownership-reconcile.ts)
plus the three subprocess-mock test harnesses now target volume-ownership.js.
