+++ packages/lib/src/control-plane/lifecycle.rollback.test.ts
@@ -20,6 +20,7 @@ const lifecycleUrl = new URL('./lifecycle.ts', import.meta.url).href;
 const moduleUrls = {
   composeArgs: new URL('./compose-args.js', import.meta.url).href,
   docker: new URL('./docker.js', import.meta.url).href,
+  volumeOwnership: new URL('./volume-ownership.js', import.meta.url).href,
   configPersistence: new URL('./config-persistence.js', import.meta.url).href,
   uiAssets: new URL('./ui-assets.js', import.meta.url).href,
   installLock: new URL('./install-lock.js', import.meta.url).href,
@@ -106,6 +107,8 @@ mock.module(${JSON.stringify(moduleUrls.docker)}, () => ({
   }),
   composeConfigServices: async () => ({ ok: true, services: [] }),
   resolveComposeProjectName: () => 'openpalm',
+}));
+mock.module(${JSON.stringify(moduleUrls.volumeOwnership)}, () => ({
   repairRootOwnedBindMounts: async () => {},
   repairManagedNamedVolumes: async () => {},
 }));
@@ -299,6 +302,8 @@ mock.module(${JSON.stringify(moduleUrls.docker)}, () => ({
   composeUp: async () => ({ ok: true, stdout: '', stderr: '', code: 0 }),
   composeConfigServices: async () => ({ ok: true, services: [] }),
   resolveComposeProjectName: () => 'openpalm',
+}));
+mock.module(${JSON.stringify(moduleUrls.volumeOwnership)}, () => ({
   repairRootOwnedBindMounts: async () => {},
   repairManagedNamedVolumes: async () => {},
 }));
@@ -425,6 +430,8 @@ mock.module(${JSON.stringify(moduleUrls.docker)}, () => ({
   composeUp: async () => ({ ok: true, stdout: '', stderr: '', code: 0 }),
   composeConfigServices: async () => ({ ok: true, services: [] }),
   resolveComposeProjectName: () => 'openpalm',
+}));
+mock.module(${JSON.stringify(moduleUrls.volumeOwnership)}, () => ({
   repairRootOwnedBindMounts: async () => {},
   repairManagedNamedVolumes: async () => {},
 }));
